tts_api/requirements.txt
# Python dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
numpy>=1.22.0
torch>=2.0.0
librosa>=0.10.0
scipy>=1.10.0
soundfile>=0.12.1
python-multipart>=0.0.6
ffmpeg-python>=0.2.0
pydantic>=2.0.0

# Text processing
inflect>=6.0.0
unidecode>=1.3.6

# Optional dependencies
#tensorboard>=2.12.0  # For training visualization
#matplotlib>=3.7.0    # For plotting